talkeasy/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
httpx
fastapi-users[sqlalchemy2]
asyncpg
python-dotenv
python-jose